Number size limit added.
The user is now warned if they enter a number too large to be
converted (which would end up creating an infinite loop and likely
crashing the page). Due to the many ways of representing decimal
numbers, there's quite a bit of code to figure out if your number
is too big.

In general, it just checks the exponent, position of the decimal,
and the value of the first several digits. If there's no exponent,
it just checks the number of digits before the decimal.