Implemented show me buttons for other bases.
All bases except decimal have a show me button, which will demonstrate
the conversion to and from that base (to or from decimal) as well as
summarize the base and its uses.

Currently the show me buttons just create a largely empty box. The
content to that will come shortly.